services/clarity_hub_api/requirements.txt
# FastAPI Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-dotenv>=1.0.0

# Supabase Integration
supabase>=2.0.0

# HTTP Requests
requests>=2.28.0
aiohttp>=3.8.0

# Caching
redis>=4.0.0

# Logging
structlog>=23.0.0

# Security
cryptography>=3.0.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
